refactor: . pull createTaskList() up
